Add http server and port listen in index.js
I am not sure if this is a right addition, but i found very annoying the fact that when i was running "crana dev:server", it wasnt exaclty clear in what port the server is listening. 
I didnt bother to add a specific framework binding like "const app = express()" - "app.listen('3000', console.log('server is running'); ", because i understand that crana is a generic way to create a node-react app with whatever node framework you want.